feat: fs watch service
