+++ packages/core/src/files/fs/file-system.ts
@@ -2,7 +2,7 @@ import { promises as fs } from 'node:fs';
 import path from 'node:path';
 import { err, ok, type Result } from '@emdash/shared';
 import { globIterate } from 'glob';
-import { realpathOrResolve } from '../../services/fs-watch';
+import { realpathOrResolve } from '../../services/fs-watch/impl/paths';
 import { enumerate as enumerateFiles } from '../enumerate';
 import { classifyFileError, isFileNotFoundCode, type FileError } from '../errors';
 import { validateAbsolutePath } from '../paths';
