+++ packages/core/src/services/fs-watch/api/index.ts
@@ -1,0 +1,18 @@
+export {
+  fsWatchContract,
+  watchErrorSchema,
+  watchEventSchema,
+  watchEventsBatchSchema,
+  watchKeySchema,
+  watchResyncSchema,
+  type FsWatchEvent,
+  type FsWatchKey,
+  type FsWatchStreamEvent,
+} from './contract';
+export type {
+  IWatchService,
+  WatchEvent,
+  WatchEventKind,
+  WatchHandle,
+  WatchOptions,
+} from './models';
