+++ packages/core/src/services/fs-watch/impl/backend.ts
@@ -1,0 +1,19 @@
+import type { Scope } from '@emdash/wire/util';
+import type { WatchEvent } from '../api';
+
+export type WatchKey = {
+  root: string;
+  ignore: string[];
+};
+
+export type WatchSink = {
+  events(events: WatchEvent[]): void;
+  resync(): void;
+};
+
+export type WatchOnError = (context: string, error: unknown) => void;
+
+export interface WatchBackend {
+  subscribe(key: WatchKey, sink: WatchSink, scope: Scope): Promise<void>;
+  dispose?(): Promise<void> | void;
+}
