+++ packages/core/src/services/fs-watch/impl/native-backend.ts
@@ -1,23 +1,5 @@
-import type { Scope } from '@emdash/wire/util';
+import type { WatchBackend, WatchOnError } from './backend';
 import { NativeWatch, type ParcelSubscribeFn } from './native-watch';
-import type { WatchEvent } from './types';
-
-export type WatchKey = {
-  root: string;
-  ignore: string[];
-};
-
-export type WatchSink = {
-  events(events: WatchEvent[]): void;
-  resync(): void;
-};
-
-export type WatchOnError = (context: string, error: unknown) => void;
-
-export interface WatchBackend {
-  subscribe(key: WatchKey, sink: WatchSink, scope: Scope): Promise<void>;
-  dispose?(): Promise<void> | void;
-}
 
 export type NativeWatchBackendOptions = {
   onError?: WatchOnError;
