+++ packages/core/src/services/fs-watch/impl/paths.ts
