--- packages/core/src/services/fs-watch/node/index.ts
@@ -1,1 +1,0 @@
-export { bootFsWatchProcess, type BootFsWatchProcessOptions } from './boot';
