+++ packages/core/src/services/fs-watch/worker/index.ts
@@ -1,0 +1,2 @@
+export { runFsWatchWorkerProcess, type RunFsWatchWorkerProcessOptions } from './process-entry';
+export { spawnFsWatchWorker, type SpawnFsWatchWorkerOptions } from './spawn';
