+++ packages/core/tsdown.config.ts
@@ -11,8 +11,8 @@ export default defineConfig({
     skills: 'src/skills/index.ts',
     exec: 'src/exec/index.ts',
     files: 'src/files/index.ts',
-    'services-fs-watch': 'src/services/fs-watch/index.ts',
-    'services-fs-watch-node': 'src/services/fs-watch/node/index.ts',
+    'services-fs-watch-api': 'src/services/fs-watch/api/index.ts',
+    'services-fs-watch-worker': 'src/services/fs-watch/worker/index.ts',
     git: 'src/git/index.ts',
     lib: 'src/lib/index.ts',
     pty: 'src/pty/index.ts',
